+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
-**./instructions.md**
-* Removed stray horizontal rule
-* Between "Parsing" and "Pathfinding" headings under "Implementation Details"
+**./src/common/graph-tasks.js**
+* Wrote new function 'getNoRouteTextString'
+* Output text for "NO SUCH ROUTE"
+
+---
+
+**./src/exact-route.js - loopSteps**
+* Negative output text retrieved with 'graphTasks.getNoRouteText'
+
+---
+
+**./src/shortest-route.js - checkRouteSuccessful**
+* Negative output text retrieved with 'graphTasks.getNoRouteText'